Minor bug fix
git-svn-id: https://ravl.svn.sourceforge.net/svnroot/ravl@3023 7226f50c-d813-0410-99bf-979b288ab488

+++ main/Contrib/ClipStationPro/CSPControl.cc
@@ -34,7 +34,7 @@ namespace RavlImageN {
       captureAudio(true), // Capture audio ?
       captureVideo(true), // Capture video ?
       frameBufferSize(0),
-      timecode_from_getframe((long int)0)
+      timecode_from_getframe((int)0)
   {
     dev = sv_open((char *) devName.chars());
     ONDEBUG(cerr << "CSP Device open:" << ((void *) dev) << " (" << devName << ")\n");
